Update patch set 4
Patch Set 4: Code-Review-1

The options structure sets up the options for an NVMe-oF target as defined by the nvmf library. However, the acceptor poll rate is an implementation detail of the nvmf application. The poll rate shouldn't be an option in that structure - the library will never use it.

Patch-set: 4
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1