Update patch set 9
Patch Set 9: Code-Review-1

(11 comments)

Please add very clear description on top of the scripts why they are separate from all other autotest.sh scripts. That their purpose is to be run on separate machines with hardware to support, instead of virtual interfaces on localhost. And so on.

Patch-set: 9
Label: Code-Review=-1