Update patch set 4
Patch Set 4:

Hi Pawel,

Thank you for explaining shutdown and its expected usage in SPDK.
But it's not possible for us to know that by the title or the comment.

Would you explain about a little more clearly? Would you add test code to validate the function?

I understand the value of smooth shutdown but I'm sorry that I'm not confident in the usage of shutdown() yet.

shutdown() may be implementation dependent.
Your intention looks that 1) stop receive, 2) send response if any, and 3) close the connection. But how do we detect the completion of 2)?

Patch-set: 4
Label: Code-Review=0