subsystem: Add an new parameter options to subsystem for initialization
Add a parameter options to struct spdk_subsystem.

This parameter is passed to init() and fini(). In init(), options is
used to initialize the subsystem. In fini(), options is freed.
When init() fails, it is responsible to free options.

This parameter is used in NVMf subsystem and iSCSI subsystem to
hold options. In other subsystems, this parameter is set to NULL.

This parameter is set by the upcoming JSON RPCs whose state is
RPC_STATE_PRE_SUBSYSTEM_INIT.

Change-Id: If7cf3a8c0a143555630feef7ebbe5da83dd0ca58
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>