Update patch set 2
Patch Set 2:

> When a poll group is created, it should automatically set up its
 > thread-local data based on the currently existing subsystems. I'm
 > not sure this is necessary.

The only reason that it is necessary is because the call to parse_conf (now parse_conf_subsystems) calls subsystem_add_ns. In the next patch, I add a call to subsystem_update_ns to subsystem_add_ns which then asynchronously calls spdk_nvmf_poll_group_update_subsystem. That function requires that the poll groups be initialized prior to calling it. In order to initialize the poll groups, the nvmf_tgt is needed which is parsed from the config file. So I parse and create the target, then create the poll groups, then parse and add the subsystems with their namespaces and update the poll group context.
This helps to avoid the overhead of adding a second add_ns function to the API. That was my initial solution to the seg faults I was getting during initialization of the NVMe-oF target.

Patch-set: 2
Reviewer: Seth Howell <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0