Update patch set 25
Patch Set 25:

I see that I previously asked for the enable and set limit calls to be split - I think that API could also make sense, but what we have now is a mix of both.

Either we should have a single API that does both enable + set limit (as my comment just above mentions), or we should have one API that enables QoS and one that sets the limit (which should support modifying the limit at runtime as well - this should encompass what the later "adjust" patch is adding in a single call).

Patch-set: 25