bdev: a new public function to get the I/O statistics of bdevs
Add a new function and its RPC caller. By using it, we can
get the statistics of all the bdevs or the specified bdev.

Meanwhile, with this patch, the open source tool 'sysstat/iostat'
can support for SPDK. The 'iostat' tool can call this function to
get the statistics of all the SPDK managed devices via the rpc
interface.

Change-Id: Iae816a0143b9641e4796e9c112f9d375f5bd9f99
Signed-off-by: Yanbo Zhou <yanbo.zhou@intel.com>