Update patch set 16
Patch Set 16:

Didn't see John's typos but note that this isn't up for that kind of review right now.  Appreciate it but there's still more work to do and comments are likely to change anyways :)

This latest patch totally reworks init code.  Previously I was creating a PMD per channel as that was optimal for virutal PMDs.  I had assumed that for QAT I could init the HW early on and have a QP per channel however upon further investigation that's not going to work.  We need to create all QP for QAT up fron based on the number of cores in the system and the number of QAT devices present, QPs are not dynamic like they are for NVMe.

So, now we init all PMDs after the bdev layer initial init is complete.  This way we know which PMDs we need to init and once they are all ready we associate them with vbdevs.  Basically all the PMD stuff is static.  This has the downside of not being able to support an RPC created vbdev that uses a virtual PMD unless there was one present when things were initilaized but that's fixable.

First, I wanted to get this framework in place for virutal PMDs present at init time as it should directly extend/support QAT as well....

Patch-set: 16