Update patch set 31
Patch Set 31:

(9 comments)

Hi Young Tack,

Thank you so much for working for this and sorry for being late for my feedback, and it was good to know you can recover the access to Gerrit.


Current config file may be deprecated in the next few releases.
when JSON config file is supported, options for DPDK will be passed by command line options (-w and -b).

Hence we should support -w and -b options in the command line as well as current config file (In SPDK these may be -W and -B respectively).

It will be great if we share the source code as much as possible between current config file and command line options.

Hence how about the following idea?

1. 
 add pci-whitelist and pci-blacklist into spdk_app_opts and spdk_env opts.

I don't think whitelist and blacklist should be mutually exclusive because DPDK's usage is not. But pci_wl/bl_addrs may not be so good, hence your idea is appreciated.

 
2.
 extracting your code int the for loop from spdk_app_setup_env() into a function. The function can be shared between read_config_file() and arg_parses().

The loop and the function can be moved from app_setup_env() to read_config_file().

Thanks,
Shuhei

Patch-set: 31