test/iscsi: allow running iSCSI tests with kernel and VPP in single run
Previously to run iSCSI tests with kernel SPDK_TEST_ISCSI=1
and SPDK_TEST_VPP=0 had to be set. Alternatively both had to
be enabled to run iSCSI test with VPP.

This patch allows to run both in single test run.

Please note that when both are enabled, VPP support is compiled in.
During subsystem initialization attempt to connect to VPP application will
be done. If no VPP application is present, 100sec timeout will be imposed for
each SPDK target start up.
Because of that when VPP is compiled into SPDK, dummy VPP app is run alongside it.

Change-Id: Ifb25a9e4b04512de882883278b8bb465de579828
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>