Update patch set 2
Patch Set 2: Code-Review-1

Thanks for looking into this.

I think we can tweak this slightly so that we don't have to submit the Set Features twice - we should be able to detect which bits are valid based on the controller's reported version and features.

For example, the NS Attribute Notices and Firmware Activation Notices bits were added in NVMe 1.2, and the Telemetry Log Notices bit was added in NVMe 1.3, so we should only set those bits if the controller version is greater than or equal to the spec version that introduced them.  Additionally, it probably doesn't make sense to set the Telemetry bit unless the controller sets cdata->lpa.telemetry.

Patch-set: 2
Label: Code-Review=-1