Update patch set 2
Patch Set 2:

Hi Ziye,

We're removing the 18.04 tag since this is not critical for 18.04 release.  There are some other implications with this patch - for example, we need to clearly state that a different env implementation *must* cache the pid, since we may now make assumptions in other SPDK code that the pid is cached.

Not saying we will not merge this, but we do not want to hold up the release for this.

-Jim

Patch-set: 2
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0