Update patch set 2
Patch Set 2: Code-Review-1

I would like to make this mutually exclusive in the script.  So keep norandommap=1 if verify is not specified.

norandommap=1 is important when doing performance testing.  It is why we added norandommap=1.  Without it, fio keeps a map of all of the blocks to keep track of which ones it has read/written so that over time, all blocks will be read/written.  This adds a lot of overhead, and you can see performance degrade.  I spent 2-3 days debugging this issue early on in the SPDK project.  :-)

Patch-set: 2
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1