[RFC]malloc_bdev: Show which copy engine is used for each channel (IOAT/memcpy)
This patch is for user to know which copy engine is used.

Subsequent patces will change the default setting of enumeration of
IOAT copy engine from enabled to disabled. The reason is to make IOAT
copy engine configurable by JSON RPC.

To make this change acceptable for user, we should provide any
method to know which copy engine is used for each channel of malloc
bdev.

Change-Id: If104f33ca0ec3a3390693f0d83189ae71beb6791
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>