iscsi: Move head of tail queues from struct iscsi_globals to static
This is a preparation to allocate iscsi_globals dynamically.

Head of target list can be initialized statically and is accessed
only in tgt_node.c. Head of initiator group list can be initialized
statically and is acccessed only in init_grp.c. Head of portal and
portal group list can be initialized statically and are accessed only
in portal_grp.c.

Hence they can be moved from iscsi_globals to a static variable, respectively.

Change-Id: I6b0983682e09199f1cd57ac48f876bf6acbf2997
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>