nvmf: Remove the qpair from the polling group
In our code, we does not remove the qpair from the polling
group. So if there are flighting I/O on such qpair,
we may face the issue reported in
https://github.com/spdk/spdk/issues/280

And this patch tries to solve this issue,
spdk_nvmf_transport_qpair_fini is already called in the correct
thread, so we do not need to use spdk_thread_send_msg function.

Change-Id: I9120a8753bbceedb12f19562aad08fec0f0052c2
Signed-off-by: Ziye Yang <ziye.yang@intel.com>