Update patch set 28
Patch Set 28: Code-Review+2

Looks reasonable to me, but definitely needs a +2 from a blobstore expert. :)

It would be nice to be able to keep the inflated clone thin provisioned and propagate unallocated clusters from the snapshot up to the clone, but that can be done as another optimization patch later on (it would need to only do that if the given clusters is unallocated all the way down the snapshot chain).

Patch-set: 28
Reviewer: Daniel Verkamp <1010525@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2