Update patch set 1
Patch Set 1:

(12 comments)

I really need to download this and put it in my editor to jump around and check various things but I know you're just looking for general comments right now to get things sorta "in the SPDK" style before getting too much more code written.

See Pawel's note on the style script and you may want to adjust your editor for things that it doesn't like (or in many cases it will just fix things for you which is nice).

This is an awesome start though!

Patch-set: 1
CC: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>