Update patch set 1
Patch Set 1:

Sorry, your first patch says that "this interface should be cleaned up so that the user does not know anything about the bdev_io except that it is the return context."

I understand if you forward nvmf protocol directly to outside of SPDK, that is the upper layer than bdev.

However bdev_io is software protocol. Unfortunately the value to conversion from bdev_io to alternative software IO protocol is not clear for me yet.

Patch-set: 1