Update patch set 1
Patch Set 1:

> (1 comment)

Yes, this is how we implement it in our "glue" layer.
The reason we need to use a ring between the submit callback and the appliance is that IO handling may take some time, and I do not want to tie up the reactor thread.
I pointed this out in the mailing list thread, and perhaps the interface needs to be changed so that bdev_user exposes bdev_user_get_io which returns an IO command (if there is one), and move the submission queue into the bdev_user module.
What do you think?

Patch-set: 1
Reviewer: shahar.salzman@kaminario.com <1014226@d5d70762-12d0-45a1-890d-524b12d3f735>