scsi: Add open semantics in scsi layer.
This patch add open semantics in scsi layer, this will
be a helper which used for hotplug handler in iscsi layer

We are trying to handle all the hotplug event in scsi layer
now, but for iscsi target, we still need make sure the logic
in iscsi layer correct, for example: for writes command, we need
wait for all active r2t done before we delete the lun. There will
be subsequent patches to handle the hotplug event in iscsi layer
base on this patch.

Change-Id: Id65d60a112ced404645b4f6480eb67807c9caf45
Signed-off-by: Cunyin Chang <cunyin.chang@intel.com>
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>