Update patch set 4
Patch Set 4:

Yes, with this difference that start_subsystem_init can be last RPC method in JSON config so together with get_rpc_methods it will be invoked as normal method on right stage. Only thing is that it should written if '-w' option is passed when save_config RPC was called.

I'm working on those two things right now. I will paste link to that patch here so you can cherrypick it or takeover if I will be unable to finish it by tommorow.

Patch-set: 4