RFC: Dynamic threading prototype
This change demonstrates a prototype of dynamic threading in
SPDK where reactors are NOT bound to lcores.

The high level idea of the design is:

Reactors are placed in a MP-MC ring. Threads are fired up
independently. The threads then pick up a reactor from the ring
and work on it for a "slice of time". The reactor is then placed
back into the ring.

A Reactor is worked upon by, one and only one thread.
But, any thread may work on the reactor.

This is a very crude initial prototype. In my next iteration,
I plan to localize the changes and use the conf. file to control
this behavior. I hope to have somewhat cleaner code by then.
Especially the part where I overload the _lcore with the reactor_id.

This change does NOT pass all the cits.

There are multiple advantages with an approach as mentioned above:

1. Apps can scale threads based on the load of the system.
2. Reactors can be "constructed" and "destroyed" based on the fan-in
   of the ports.
3. It maintains the core lockless design philosophy and would have
   minimal churn in the core library.
4. To maintain "backward" compatibility, one could create a 1-1
   relationship between the reactors and the lcores thus allowing a
   thread to infintely run on a reactor.

There are some disadvantages with the above appoach:

1. The first one that comes to mind is losing the CPU cache benefits.
2. The event pool is also not split up between the sockets (although
   in a productized version, this could be solved).
3. The timers in the reactor rings may not work in a hard bound fashion.

Change-Id: I9baedda49c92edc27ac4e296db63aa5a665e9bf0
Signed-off-by: John Meneghini <johnm@netapp.com>
Signed-off-by: Madhu Pai <mpai@netapp.com>