Update patch set 1
Patch Set 1: Code-Review-1

I'm not sure if I understand what this is trying to fix; we already support zero-length payloads in the NVMe library.

The changes look like they are poking around in the CDW<n> fields inside the command, which are supposed to be filled out by the caller, not interpreted by the transport (nvme_pcie in this case); is this some kind of OpenChannel-specific workaround?

Patch-set: 1
Reviewer: Daniel Verkamp <1010525@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1