Update patch set 47
Patch Set 47:

We are going to put more test into this script(nvmf, iscsi), and also test parameters for spdk_tgt. I think that we are going to test only rpc with this script so I think that this place is appropriate for json-rpc tests. We are not going to use fio to test bdevs. To be precise this scripts is going to test save_config, load_config rpc methods. Other use of rpc is because we need to create full environment/settings of spdk_tgt to test save_config, load_config.

Patch-set: 47
Reviewer: Pawel Kaminski <1011260@d5d70762-12d0-45a1-890d-524b12d3f735>