Update patch set 48
Patch Set 48:

(1 comment)

I'm OK with keeping this as a standalone "JSON config" test script if the intent is to add tests for all components to it.

However, I think the Jenkins test pool infrastructure is set up so that each machine runs a particular slice of the tests, and we don't want to have nested feature tests (e.g. tests that check both iSCSI + JSON test flags).  Karol, any comments from the Jenkins point of view?

Patch-set: 48