Update patch set 32
Patch Set 32: Code-Review-1

This will freeze all I/O to the entire blobstore - not just the blob that is being snapshotted (or in the next patch, resized).

The flag needs to be on the blob, not the channel.  Set the flag, do an empty message to each channel, and then you can proceed with the snapshot, resize, etc. operation.

There's more to it then just this - but please think through this some more and do your best to come up with a more complete solution.

Thanks,

-Jim

Patch-set: 32
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1